refactor(int4-gemm): consolidate f32_to_bf16_bits into single source-of-truth
PR #30 production-readiness review SHOULD-FIX #1. Four identical
copies of the f32->bf16 round-to-nearest-even helper drifted into:

- tahoma-engine-sparse-moe/src/runner.rs::f32_to_bf16_bits
  (the only one with a cross-check test against half::bf16::from_f32)
- tahoma-int4-gemm/src/shell_int4.rs::f32_to_bf16_bits_local
- tahoma-int4-gemm/src/layer0_int4.rs::f32_to_bf16_bits_local
- tahoma-int4-gemm/src/c_ffi.rs::f32_to_bf16_bits_ffi

All four were bit-identical implementations. Pick tahoma-int4-gemm's
format.rs as the canonical home — it already houses the matching
bf16_bits_to_f32 helper and the bf16-as-u16 storage layout docs.

Move:
- The function (now pub) -> format.rs as f32_to_bf16_bits.
- The half::bf16 cross-check test -> format::tests, pinning the
  contract once for all callers.
- Re-export from tahoma-int4-gemm's lib.rs so the sparse-moe runner
  can import without referencing format::.

The four call-sites now all route through this one function. No
behavior change — same byte output, same NaN canonicalization, same
rounding mode. Verified by running `cargo test -p tahoma-int4-gemm`
(23 tests pass including the cross-check) and `cargo test -p
tahoma-engine-sparse-moe` (62 tests across all targets pass).
