+++ crates/tahoma-int4-gemm/src/lib.rs
@@ -38,7 +38,7 @@ pub mod safetensors_source;
 pub mod shell;
 pub mod shell_int4;
 
-pub use format::{ExpertWeights, GemmError};
+pub use format::{bf16_bits_to_f32, f32_to_bf16_bits, ExpertWeights, GemmError};
 pub use kernel::{dequant_gemv_int4, expert_forward};
 pub use kernel_avx512::dequant_gemv_int4_auto;
 pub use kernel_avx512_multi::dequant_gemm_int4_multi_auto;
